+++ demisto_sdk/commands/common/tests/tools_core_test.py
@@ -2,28 +2,28 @@ from pathlib import Path
 
 import pytest
 
-from demisto_sdk.commands.common.tools_core import find_pack_folder, string_to_bool
+from demisto_sdk.commands.common import tools_core
 
 
 @pytest.mark.parametrize("value", ("true", "True", 1, "1", "yes", "y"))
 def test_string_to_bool_true(value: str):
-    assert string_to_bool(value)
+    assert tools_core.string_to_bool(value)
 
 
 @pytest.mark.parametrize("value", ("", None))
 def test_string_to_bool_default_true(value: str):
-    assert string_to_bool(value, True)
+    assert tools_core.string_to_bool(value, True)
 
 
 @pytest.mark.parametrize("value", ("false", "False", 0, "0", "n", "no"))
 def test_string_to_bool_false(value: str):
-    assert not string_to_bool(value)
+    assert not tools_core.string_to_bool(value)
 
 
 @pytest.mark.parametrize("value", ("", " ", "כן", None, "None"))
 def test_string_to_bool_error(value: str):
     with pytest.raises(ValueError):
-        string_to_bool(value)
+        tools_core.string_to_bool(value)
 
 
 @pytest.mark.parametrize(
@@ -40,5 +40,29 @@ def test_string_to_bool_error(value: str):
     ],
 )
 def test_find_pack_folder(input_path, expected_output):
-    output = find_pack_folder(input_path)
+    output = tools_core.find_pack_folder(input_path)
     assert expected_output == str(output)
+
+
+@pytest.mark.parametrize(
+    "input_path, expected_output",
+    [
+        (
+            Path(
+                "/User/username/content/Packs/MyPack/Integrations/MyIntegration/MyIntegration.yml"
+            ),
+            Path("/User/username/content"),
+        ),
+        (Path("/User/username/content/Packs"), Path("/User/username/content")),
+    ],
+)
+def test_get_content_path(input_path, expected_output):
+    """
+    Given:
+        - A path to a file or directory in the content repo
+    When:
+        - Running get_content_path
+    Then:
+        Validate that the given path is correct
+    """
+    assert tools_core.get_content_path(input_path) == expected_output
