+++ demisto_sdk/commands/common/tests/tools_test.py
@@ -3009,30 +3009,6 @@ def test_sha1_dir():
 
 
 @pytest.mark.parametrize(
-    "input_path, expected_output",
-    [
-        (
-            Path(
-                "/User/username/content/Packs/MyPack/Integrations/MyIntegration/MyIntegration.yml"
-            ),
-            Path("/User/username/content"),
-        ),
-        (Path("/User/username/content/Packs"), Path("/User/username/content")),
-    ],
-)
-def test_get_content_path(input_path, expected_output):
-    """
-    Given:
-        - A path to a file or directory in the content repo
-    When:
-        - Running get_content_path
-    Then:
-        Validate that the given path is correct
-    """
-    assert tools.get_content_path(input_path) == expected_output
-
-
-@pytest.mark.parametrize(
     "string, expected_result",
     [
         ("1", True),
